refactor(pipeline/api-direct-call): extract SignerTypes from IApiDirectCallConfig
Phase 8 commit 2/8 of ApiDirectCallConfig split.

- Carve SignerTypes (SignerAlgorithm, AsymmetricSignerAlgorithm,
  SignerEncoding, CanonicalPart, ICanonicalStringConfig,
  IAsymmetricSignerConfig, IAesSignerConfig, ISignerConfig,
  ICryptoFieldConfig) from
  src/Scrapers/Pipeline/Mediator/ApiDirectCall/IApiDirectCallConfig.ts.
- New file:
  src/Scrapers/Pipeline/Mediator/ApiDirectCall/ConfigContracts/SignerTypes.ts
  (101 raw LoC; no intra-cluster deps — bucket-1 standalone).
- Comfortably under the canonical 150-LoC cap; no sub-split needed
  despite spec.txt §1 estimating ~160 LoC.
- IApiDirectCallConfig.ts now imports the nine types and continues to
  re-export them from its existing `export type {}` block so all 53
  callers compile unchanged.

Verified:
- npm run type-check green
- npx tsup green (lib/index.cjs byte-identical to main)
- npx jest --testPathPatterns=ApiDirectCall green (223/223)
- New sub-module 101 LoC (under 150 cap)

Master plan: pipeline-decoupling-master-2026-05-28 / phase-8
